+++ packages/isar_connect/lib/isar_connect.dart
@@ -1,3 +1,0 @@
-library isar_connect;
-
-export 'src/isar_connect_base.dart' show initializeIsarConnect;